Adds basic iOS C examples
Adds two basic examples for iOS written in C:
- A simple Metal triangle rendering example
- A touch input example that draws a rectangle on touch

These examples are part of the RGFW framework.
They showcase how to create a window and attach it to the first scene,
and how to handle touch events.

The iOS examples use Metal for rendering.

Also updates premake to correctly build the examples.